chore: removed libft_bonus.h since it ain't needed
