+++ libft/ft_lstclear_bonus.c
@@ -6,11 +6,11 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/11 08:25:26 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/13 20:14:34 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/14 07:34:51 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_bonus.h"
+#include "libft.h"
 
 void	ft_lstclear(t_list **lst, void (*del)(void *))
 {
