+++ libft/ft_lstiter_bonus.c
@@ -6,11 +6,11 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/11 08:55:27 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/13 20:14:51 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/14 07:35:01 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_bonus.h"
+#include "libft.h"
 
 void	ft_lstiter(t_list *lst, void (*f)(void *))
 {
