+++ libft/ft_lstmap_bonus.c
@@ -6,11 +6,11 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/11 08:57:44 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/13 20:15:00 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/14 07:35:10 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_bonus.h"
+#include "libft.h"
 
 /**
  * Applies the function 'f' to each element of the linked list 'lst'
