+++ libft/ft_lstnew_bonus.c
@@ -6,11 +6,11 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/10 20:02:01 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/13 20:15:04 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/14 07:35:16 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_bonus.h"
+#include "libft.h"
 
 t_list	*ft_lstnew(void *content)
 {
