+++ libft/ft_lstsize_bonus.c
@@ -6,11 +6,11 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/10 20:42:32 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/13 20:27:09 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/14 07:35:21 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft_bonus.h"
+#include "libft.h"
 
 int	ft_lstsize(t_list *lst)
 {
